move commands into pkg/
