+++ cmd/kind/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"sigs.k8s.io/kind/cmd/kind/app"
+)
+
+func main() {
+	app.Main()
+}
