+++ pkg/cmd/kind/completion/bash/bash.go
