+++ pkg/cmd/kind/delete/cluster/deletecluster.go
