+++ pkg/cmd/kind/get/nodes/nodes.go
