+++ pkg/cmd/kind/load/docker-image/docker-image.go
